Updated Create PQ routing, by routing to the View instead of the component and renaming the component. Also changed the storage of the PartyQuest Data
