--- src/views/CreatePQPage.vue
@@ -1,14 +1,0 @@
-<template>
-    <CreatePQ />
-</template>
-
-<script>
-import CreatePQ from '../components/CreatePQ.vue';
-
-export default {
-    name: 'CreatePQ',
-    components: {
-        CreatePQ,
-    },
-};
-</script>
